+++ packages/rstream-dot/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.2.31"></a>
+## [0.2.31](https://github.com/thi-ng/umbrella/compare/@thi.ng/rstream-dot@0.2.30...@thi.ng/rstream-dot@0.2.31) (2018-08-01)
+
+
+
+
+**Note:** Version bump only for package @thi.ng/rstream-dot
+
 <a name="0.2.30"></a>
 ## [0.2.30](https://github.com/thi-ng/umbrella/compare/@thi.ng/rstream-dot@0.2.29...@thi.ng/rstream-dot@0.2.30) (2018-08-01)